Add recovery function for Alice
